fix: partial rename to vendor
