+++ src/mpflash/tests/test_mp_board_filter.py
@@ -3,7 +3,7 @@ from typing import List
 import pytest
 
 from mpflash.mpboard_id.api import get_stored_boards_for_port, local_mp_ports
-from mpflash.vendored.versions import get_stable_mp_version
+from mpflash.vendor.versions import get_stable_mp_version
 
 pytestmark = [pytest.mark.mpflash]
